+++ openmls/src/group/core_group/proposals.rs
@@ -512,6 +512,12 @@ impl ProposalQueue {
                     proposal_pool.insert(queued_proposal.proposal_reference(), queued_proposal);
                 }
                 Proposal::AppAck(_) => unimplemented!("See #291"),
+                Proposal::Custom(_) => {
+                    // Other/unknown proposals are always considered valid and
+                    // have to be checked by the application instead.
+                    valid_proposals.add(queued_proposal.proposal_reference());
+                    proposal_pool.insert(queued_proposal.proposal_reference(), queued_proposal);
+                }
             }
         }
         // Check for presence of Removes and delete Updates