+++ openmls/src/schedule/mod.rs
@@ -940,7 +940,10 @@ fn ciphertext_sample(ciphersuite: Ciphersuite, ciphertext: &[u8]) -> &[u8] {
 /// A key that can be used to derive an `AeadKey` and an `AeadNonce`.
 #[derive(Serialize, Deserialize)]
 #[cfg_attr(test, derive(PartialEq))]
-#[cfg_attr(any(feature = "test-utils", test), derive(Debug, Clone))]
+#[cfg_attr(
+    any(feature = "test-utils", feature = "crypto-debug", test),
+    derive(Debug, Clone)
+)]
 pub(crate) struct SenderDataSecret {
     secret: Secret,
 }